ml-service/requirements.txt

fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.0


httpx==0.27.0
aiohttp==3.9.3


torch==2.2.1
transformers==4.38.2
sentencepiece==0.2.0


spacy==3.7.4


numpy==1.26.4


python-dotenv==1.0.1